Refactor marks processing and export logic to improve performance and clarity. Update the ProcessMarksGrades command to compute totals and averages more efficiently, and enhance the MarksExport and MarksUserExport classes to utilize a grading facade for better grade assignment. Remove deprecated methods and streamline data handling for improved maintainability.

+++ database/migrations/2025_02_28_131137_create_mark_grades_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('mark_grades', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('markId'); // Must match marks.markId type
+            $table->unsignedBigInteger('markId'); // Must match marks.markId type ($table->id('markId') = bigint unsigned)
             $table->string('subject');
             $table->string('grade');
             $table->timestamps();
